+++ crud/options.go
@@ -23,6 +23,7 @@ const (
 	batchSizeOptName                     = "batch_size"
 	fetchLatestMetadataOptName           = "fetch_latest_metadata"
 	noreturnOptName                      = "noreturn"
+	yieldEvery                           = "yield_every"
 )
 
 // OptUint is an optional uint.